refactor(engineer_worktree): extract cleanup helpers — mod.rs <400 (#1266)
Final extraction: cleanup_inner + assert_under_root + create_worktrees_root
+ unique_suffix moved to sibling cleanup.rs.

mod.rs: 436 -> 319 (UNDER 400 LOC ✓)
cleanup.rs: 133 LOC

Together with #1366 (claim.rs) and #1367 (sweep.rs), engineer_worktree/mod.rs
is now compliant with the 400-LOC ceiling for #1266.

All 17 engineer_worktree tests pass.
